% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bitcoin_gold_oil}
\alias{bitcoin_gold_oil}
\title{bitcoin_gold_oil data set}
\format{
A data frame with 18 columns and 1827 rows.
}
\source{
Yahoo Finance
}
\usage{
bitcoin_gold_oil
}
\description{
A data frame with different time series (prices and volumes) for bitcoin, gold and oil.
}
\keyword{datasets}
